+++ src/main/java/cmanager/okapi/Okapi.java
@@ -1,7 +1,6 @@
 package cmanager.okapi;
 
 import cmanager.exception.CoordinateUnparsableException;
-import cmanager.exception.MalFormedException;
 import cmanager.exception.UnexpectedLogStatus;
 import cmanager.geo.Coordinate;
 import cmanager.geo.Geocache;
@@ -27,27 +26,32 @@ import com.github.scribejava.core.model.Verb;
 import com.github.scribejava.core.oauth.OAuth10aService;
 import com.google.gson.Gson;
 import java.io.IOException;
-import java.net.URLEncoder;
+import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 
+/** Handle communication with the OKAPI. */
 public class Okapi {
 
+    /** The HTTP client to use. */
     private static final ApacheHttp httpClient = new ApacheHttp();
 
-    public static String usernameToUuid(String username) throws Exception {
-        final String url =
-                OcSite.getBaseUrl()
-                        + "okapi/services"
-                        + "/users/by_username"
-                        + "?consumer_key="
-                        + OcSite.getConsumerApiKey()
-                        + "&username="
-                        + URLEncoder.encode(username, "UTF-8")
-                        + "&fields=uuid";
+    /**
+     * Convert the given username to an user UUID.
+     *
+     * @param username The username to convert.
+     * @return The user UUID for the given username.
+     * @throws IOException Something went wrong with the network request.
+     * @throws UnexpectedStatusCode Something went wrong with our OKAPI request.
+     * @throws URISyntaxException The URL is invalid. In theory this should never be the case for
+     *     us.
+     */
+    public static String usernameToUuid(String username)
+            throws IOException, UnexpectedStatusCode, URISyntaxException {
+        final String url = OkapiUrlBuilder.getUsernameToUuidUrl(username);
 
         final HttpResponse httpResponse = httpClient.get(url);
         final String responseBody = httpResponse.getBody();
@@ -65,19 +69,20 @@ public class Okapi {
         return document.getUuid();
     }
 
-    public static Geocache getCache(String code) throws Exception {
-        final String url =
-                OcSite.getBaseUrl()
-                        + "okapi/services"
-                        + "/caches/geocache"
-                        + "?consumer_key="
-                        + OcSite.getConsumerApiKey()
-                        + "&cache_code="
-                        + code
-                        + "&fields="
-                        + URLEncoder.encode(
-                                "code|name|location|type|gc_code|difficulty|terrain|status",
-                                "UTF-8");
+    /**
+     * Get the basic cache details for the given OC code.
+     *
+     * @param code The OC code to get the basic cache details for.
+     * @return The geocache instance with the basic cache details.
+     * @throws IOException Something went wrong with the network request.
+     * @throws UnexpectedStatusCode Something went wrong with our OKAPI request.
+     * @throws URISyntaxException The URL is invalid. In theory this should never be the case for
+     *     us.
+     * @throws NumberFormatException The coordinates could not be parsed.
+     */
+    public static Geocache getCache(String code)
+            throws IOException, UnexpectedStatusCode, URISyntaxException, NumberFormatException {
+        final String url = OkapiUrlBuilder.getCacheUrl(code);
 
         final HttpResponse httpResponse = httpClient.get(url);
         final String responseBody = httpResponse.getBody();
@@ -133,8 +138,22 @@ public class Okapi {
         return geocache;
     }
 
+    /**
+     * Get the opencache instance referenced by its OC code. If possible, the object will be
+     * returned from the runtime cache instead of requesting it again over the network.
+     *
+     * @param code The OC code to get the object for.
+     * @param okapiRuntimeCache The runtime cache of the OKAPI holding all the opencache objects of
+     *     the current application run.
+     * @return The requested opencache instance.
+     * @throws IOException Something went wrong with the network request.
+     * @throws UnexpectedStatusCode Something went wrong with our OKAPI request.
+     * @throws URISyntaxException The URL is invalid. In theory this should never be the case for
+     *     us.
+     * @throws NumberFormatException The geocache coordinates could not be parsed.
+     */
     public static Geocache getCacheBuffered(String code, List<Geocache> okapiRuntimeCache)
-            throws Exception {
+            throws IOException, UnexpectedStatusCode, URISyntaxException, NumberFormatException {
         synchronized (okapiRuntimeCache) {
             final int index = Collections.binarySearch(okapiRuntimeCache, code);
             if (index >= 0) {
@@ -152,20 +171,19 @@ public class Okapi {
         return geocache;
     }
 
-    public static Geocache completeCacheDetails(Geocache geocache) throws Exception {
-        final String url =
-                OcSite.getBaseUrl()
-                        + "okapi/services"
-                        + "/caches/geocache"
-                        + "?consumer_key="
-                        + OcSite.getConsumerApiKey()
-                        + "&cache_code="
-                        + geocache.getCode()
-                        + "&fields="
-                        + URLEncoder.encode(
-                                "size2|short_description|description|owner|hint2|url|req_passwd"
-                                        + "|internal_id",
-                                "UTF-8");
+    /**
+     * Complete the details for the given opencache instance.
+     *
+     * @param geocache The opencache instance to complete the cache details for.
+     * @return The opencache instance with the details added.
+     * @throws IOException Something went wrong with the network request.
+     * @throws UnexpectedStatusCode Something went wrong with our OKAPI request.
+     * @throws URISyntaxException The URL is invalid. In theory this should never be the case for
+     *     us.
+     */
+    public static Geocache completeCacheDetails(Geocache geocache)
+            throws IOException, UnexpectedStatusCode, URISyntaxException {
+        final String url = OkapiUrlBuilder.getCompleteCacheDetailsUrl(geocache);
 
         final HttpResponse httpResponse = httpClient.get(url);
         final String responseBody = httpResponse.getBody();
@@ -193,12 +211,26 @@ public class Okapi {
         return geocache;
     }
 
+    /**
+     * Build the OAuth service for the OKAPI.
+     *
+     * @return The OAuth service to use with the OKAPI.
+     */
     private static OAuth10aService getOAuthService() {
         return new ServiceBuilder(OcSite.getConsumerApiKey())
                 .apiSecret(OcSite.getConsumerSecretKey())
                 .build(new OAuth());
     }
 
+    /**
+     * Request OAuth authorization
+     *
+     * @param callback The callback interface to use.
+     * @return The OAuth access token instance to use with the requests further on.
+     * @throws IOException Something went wrong with the network request.
+     * @throws InterruptedException The execution has been interrupted.
+     * @throws ExecutionException The result of the interrupted/aborted task should be accessed.
+     */
     public static OAuth1AccessToken requestAuthorization(
             RequestAuthorizationCallbackInterface callback)
             throws IOException, InterruptedException, ExecutionException {
@@ -221,24 +253,54 @@ public class Okapi {
         return service.getAccessToken(requestToken, pin);
     }
 
+    /**
+     * Perform an HTTP GET request using OAuth.
+     *
+     * @param tokenProvider The OAuth token provider to use.
+     * @param url The URL to request.
+     * @return The result of the GET request.
+     * @throws IOException Something went wrong with the network request.
+     * @throws InterruptedException The execution has been interrupted.
+     * @throws ExecutionException The result of the interrupted/aborted task should be accessed.
+     */
     private static String authedHttpGet(
             final TokenProviderInterface tokenProvider, final String url)
             throws InterruptedException, ExecutionException, IOException {
         final OAuth10aService service = getOAuthService();
         final OAuthRequest request = new OAuthRequest(Verb.GET, url);
+
         // The access token from step 4.
         service.signRequest(tokenProvider.getOkapiToken(), request);
+
         final Response response = service.execute(request);
         return response.getBody();
     }
 
+    /**
+     * Get the opencaches around the center given by the geocache instance.
+     *
+     * @param tokenProvider The OAuth token provider to use.
+     * @param excludeUuid Set to an user UUID to exclude caches ignored by this user. If this is
+     *     set, the token provider cannot be null.
+     * @param geocache The geocache to use the coordinates from as the search center.
+     * @param searchRadius The search radius to use.
+     * @param okapiRuntimeCache The runtime cache to use for faster access to opencache instances.
+     * @return The list of opencaches around the given geocache.
+     * @throws IOException Something went wrong with the network request.
+     * @throws InterruptedException The execution has been interrupted.
+     * @throws ExecutionException The result of the interrupted/aborted task should be accessed.
+     * @throws UnexpectedStatusCode Something went wrong with our OKAPI request.
+     * @throws URISyntaxException The URL is invalid. In theory this should never be the case for
+     *     us.
+     */
     public static List<Geocache> getCachesAround(
             TokenProviderInterface tokenProvider,
             String excludeUuid,
             Geocache geocache,
             double searchRadius,
             List<Geocache> okapiRuntimeCache)
-            throws Exception {
+            throws IOException, UnexpectedStatusCode, URISyntaxException, InterruptedException,
+                    ExecutionException {
         final Coordinate coordinate = geocache.getCoordinate();
         return getCachesAround(
                 tokenProvider,
@@ -249,6 +311,25 @@ public class Okapi {
                 okapiRuntimeCache);
     }
 
+    /**
+     * Get the opencaches around the given center.
+     *
+     * @param tokenProvider The OAuth token provider to use.
+     * @param excludeUuid Set to an user UUID to exclude caches ignored by this user. If this is
+     *     set, the token provider cannot be null.
+     * @param latitude The latitude of the search center.
+     * @param longitude The longitude of the search center.
+     * @param searchRadius The search radius to use.
+     * @param okapiCacheDetailsCache The runtime cache to use for faster access to opencache
+     *     instances.
+     * @return The list of opencaches around the given geocache.
+     * @throws IOException Something went wrong with the network request.
+     * @throws InterruptedException The execution has been interrupted.
+     * @throws ExecutionException The result of the interrupted/aborted task should be accessed.
+     * @throws UnexpectedStatusCode Something went wrong with our OKAPI request.
+     * @throws URISyntaxException The URL is invalid. In theory this should never be the case for
+     *     us.
+     */
     public static List<Geocache> getCachesAround(
             TokenProviderInterface tokenProvider,
             String excludeUuid,
@@ -256,24 +337,12 @@ public class Okapi {
             Double longitude,
             Double searchRadius,
             List<Geocache> okapiCacheDetailsCache)
-            throws Exception {
+            throws IOException, UnexpectedStatusCode, URISyntaxException, InterruptedException,
+                    ExecutionException {
         final boolean useOAuth = tokenProvider != null && excludeUuid != null;
         final String url =
-                OcSite.getBaseUrl()
-                        + "okapi/services"
-                        + "/caches/search/nearest"
-                        + "?consumer_key="
-                        + OcSite.getConsumerApiKey()
-                        + "&center="
-                        + URLEncoder.encode(
-                                latitude.toString() + "|" + longitude.toString(), "UTF-8")
-                        + "&radius="
-                        + searchRadius.toString()
-                        + "&status="
-                        + URLEncoder.encode("Available|Temporarily unavailable|Archived", "UTF-8")
-                        + "&limit=500"
-                        + (useOAuth ? "&ignored_status=notignored_only" : "")
-                        + (useOAuth ? "&not_found_by=" + excludeUuid : "");
+                OkapiUrlBuilder.getCachesAroundUrl(
+                        useOAuth, excludeUuid, latitude, longitude, searchRadius);
 
         String responseBody;
         if (useOAuth) {
@@ -300,38 +369,52 @@ public class Okapi {
                 if (geocache != null) {
                     caches.add(geocache);
                 }
-            } catch (MalFormedException exception) {
+            } catch (NumberFormatException exception) {
                 ExceptionPanel.display(exception);
             }
         }
         return caches;
     }
 
-    public static void updateFoundStatus(TokenProviderInterface tokenProvider, Geocache oc)
-            throws IOException, InterruptedException, ExecutionException {
+    /**
+     * Retrieve the found status of the given opencache instance and update the object accordingly.
+     *
+     * @param tokenProvider The OAuth token provider to use.
+     * @param opencache The opencache to search for and to update.
+     * @throws IOException Something went wrong with the network request.
+     * @throws InterruptedException The execution has been interrupted.
+     * @throws ExecutionException The result of the interrupted/aborted task should be accessed.
+     * @throws URISyntaxException The URL is invalid. In theory this should never be the case for
+     *     us.
+     */
+    public static void updateFoundStatus(TokenProviderInterface tokenProvider, Geocache opencache)
+            throws IOException, InterruptedException, ExecutionException, URISyntaxException {
         if (tokenProvider == null) {
             return;
         }
 
-        final String url =
-                OcSite.getBaseUrl()
-                        + "okapi/services"
-                        + "/caches/geocache"
-                        + "?consumer_key="
-                        + OcSite.getConsumerApiKey()
-                        + "&cache_code="
-                        + oc.getCode()
-                        + "&fields=is_found";
+        final String url = OkapiUrlBuilder.getFoundStatusUrl(opencache);
         final String responseBody = authedHttpGet(tokenProvider, url);
 
         final GeocacheDocument document = new Gson().fromJson(responseBody, GeocacheDocument.class);
 
-        oc.setIsFound(document.isFound());
+        opencache.setIsFound(document.isFound());
     }
 
+    /**
+     * Get the UUID of the currently authenticated user.
+     *
+     * @param tokenProvider The OAuth token provider to use.
+     * @return The UUID of the currently authenticated user.
+     * @throws IOException Something went wrong with the network request.
+     * @throws InterruptedException The execution has been interrupted.
+     * @throws ExecutionException The result of the interrupted/aborted task should be accessed.
+     * @throws URISyntaxException The URL is invalid. In theory this should never be the case for
+     *     us.
+     */
     public static String getUuid(TokenProviderInterface tokenProvider)
-            throws IOException, InterruptedException, ExecutionException {
-        final String url = OcSite.getBaseUrl() + "okapi/services" + "/users/user" + "?fields=uuid";
+            throws IOException, InterruptedException, ExecutionException, URISyntaxException {
+        final String url = OkapiUrlBuilder.getUuidUrl();
         final String responseBody = authedHttpGet(tokenProvider, url);
 
         final UserDocument document = new Gson().fromJson(responseBody, UserDocument.class);
@@ -341,10 +424,20 @@ public class Okapi {
         return document.getUuid();
     }
 
+    /**
+     * Get the username of the currently authenticated user.
+     *
+     * @param tokenProvider The OAuth token provider to use.
+     * @return The username of the currently authenticated user.
+     * @throws IOException Something went wrong with the network request.
+     * @throws InterruptedException The execution has been interrupted.
+     * @throws ExecutionException The result of the interrupted/aborted task should be accessed.
+     * @throws URISyntaxException The URL is invalid. In theory this should never be the case for
+     *     us.
+     */
     public static String getUsername(TokenProviderInterface tokenProvider)
-            throws IOException, InterruptedException, ExecutionException {
-        final String url =
-                OcSite.getBaseUrl() + "okapi/services" + "/users/user" + "?fields=username";
+            throws IOException, InterruptedException, ExecutionException, URISyntaxException {
+        final String url = OkapiUrlBuilder.getUsernameUrl();
         final String responseBody = authedHttpGet(tokenProvider, url);
 
         final UserDocument document = new Gson().fromJson(responseBody, UserDocument.class);
@@ -354,29 +447,32 @@ public class Okapi {
         return document.getUsername();
     }
 
+    /**
+     * Post the log for the given opencache.
+     *
+     * @param tokenProvider The OAuth token provider to use.
+     * @param cache The opencache to post the log for.
+     * @param log The log to post.
+     * @param returnInternalIdInsteadOfUuid If this is true, the internal log ID will be returned
+     *     instead of the log UUID.
+     * @return Either the internal or the UUID of the posted log, depending on the boolean variable.
+     * @throws IOException Something went wrong with the network request.
+     * @throws InterruptedException The execution has been interrupted.
+     * @throws ExecutionException The result of the interrupted/aborted task should be accessed.
+     * @throws UnexpectedStatusCode Something went wrong with our OKAPI request.
+     * @throws URISyntaxException The URL is invalid. In theory this should never be the case for
+     *     us.
+     * @throws UnexpectedLogStatus Posting the log has not been successful according to the OKAPI
+     *     response.
+     */
     public static String postLog(
             TokenProviderInterface tokenProvider,
             Geocache cache,
             GeocacheLog log,
             boolean returnInternalIdInsteadOfUuid)
             throws InterruptedException, ExecutionException, IOException, UnexpectedLogStatus,
-                    UnexpectedStatusCode {
-        String url =
-                OcSite.getBaseUrl()
-                        + "okapi/services"
-                        + "/logs/submit"
-                        + "?cache_code="
-                        + URLEncoder.encode(cache.getCode(), "UTF-8")
-                        + "&logtype="
-                        + URLEncoder.encode(log.getOkapiType(cache), "UTF-8")
-                        + "&comment="
-                        + URLEncoder.encode(log.getText(), "UTF-8")
-                        + "&when="
-                        + URLEncoder.encode(log.getDateStrIso8601NoTime(), "UTF-8");
-
-        /*if (cache.doesRequirePassword() != null && cache.doesRequirePassword()) {
-            url += "&password=" + URLEncoder.encode(log.getPassword(), "UTF-8");
-        }*/
+                    UnexpectedStatusCode, URISyntaxException {
+        String url = OkapiUrlBuilder.getLogSubmissionUrl(cache, log);
 
         final String responseBody = authedHttpGet(tokenProvider, url);
 
@@ -405,18 +501,24 @@ public class Okapi {
         return document.getLogUuid();
     }
 
+    /**
+     * Get the home coordinates for the currently authenticated user.
+     *
+     * @param tokenProvider The OAuth token provider to use.
+     * @return The home coordinates of the currently authenticated user.
+     * @throws IOException Something went wrong with the network request.
+     * @throws InterruptedException The execution has been interrupted.
+     * @throws ExecutionException The result of the interrupted/aborted task should be accessed.
+     * @throws URISyntaxException The URL is invalid. In theory this should never be the case for
+     *     us.
+     * @throws CoordinateUnparsableException The coordinates could not be parsed.
+     */
     public static Coordinate getHomeCoordinates(TokenProviderInterface tokenProvider)
             throws CoordinateUnparsableException, IOException, InterruptedException,
-                    ExecutionException {
+                    ExecutionException, URISyntaxException {
         final String uuid = getUuid(tokenProvider);
 
-        final String url =
-                OcSite.getBaseUrl()
-                        + "okapi/services"
-                        + "/users/user"
-                        + "?fields=home_location"
-                        + "&user_uuid="
-                        + uuid;
+        final String url = OkapiUrlBuilder.getHomeCoordinatesUrl(uuid);
         final String responseBody = authedHttpGet(tokenProvider, url);
 
         final UserDocument document = new Gson().fromJson(responseBody, UserDocument.class);
@@ -424,16 +526,19 @@ public class Okapi {
         return document.getHomeLocationAsCoordinate();
     }
 
-    /** Convert the given log UUID to a real (internal) log ID. */
-    public static String getLogId(String logUuid) throws IOException, UnexpectedStatusCode {
-        final String url =
-                OcSite.getBaseUrl()
-                        + "okapi/services"
-                        + "/logs/entry?consumer_key="
-                        + OcSite.getConsumerApiKey()
-                        + "&fields=internal_id"
-                        + "&log_uuid="
-                        + URLEncoder.encode(logUuid, "UTF-8");
+    /**
+     * Convert the given log UUID to a real (internal) log ID.
+     *
+     * @param logUuid The log UUID to convert.
+     * @return The real (internal) log ID for the given log UUID.
+     * @throws IOException Something went wrong with the network request.
+     * @throws UnexpectedStatusCode Something went wrong with our OKAPI request.
+     * @throws URISyntaxException The URL is invalid. In theory this should never be the case for
+     *     us.
+     */
+    public static String getLogId(String logUuid)
+            throws IOException, UnexpectedStatusCode, URISyntaxException {
+        final String url = OkapiUrlBuilder.getLogIdUrl(logUuid);
 
         final HttpResponse httpResponse = httpClient.get(url);
         final String responseBody = httpResponse.getBody();
@@ -451,15 +556,21 @@ public class Okapi {
      *
      * <p>This is only needed for the automated tests. It will probably be never implemented in the
      * GUI (and it probably should not be implemented there anyway).
+     *
+     * @param tokenProvider The OAuth token provider to use.
+     * @param logUuid The UUID of the log to delete.
+     * @throws IOException Something went wrong with the network request.
+     * @throws InterruptedException The execution has been interrupted.
+     * @throws ExecutionException The result of the interrupted/aborted task should be accessed.
+     * @throws URISyntaxException The URL is invalid. In theory this should never be the case for
+     *     us.
+     * @throws UnexpectedLogStatus Deleting the log has not been successful according to the OKAPI
+     *     response.
      */
     public static void deleteLog(TokenProviderInterface tokenProvider, String logUuid)
-            throws IOException, ExecutionException, InterruptedException, UnexpectedLogStatus {
-        String url =
-                OcSite.getBaseUrl()
-                        + "okapi/services"
-                        + "/logs/delete"
-                        + "?log_uuid="
-                        + URLEncoder.encode(logUuid, "UTF-8");
+            throws IOException, ExecutionException, InterruptedException, UnexpectedLogStatus,
+                    URISyntaxException {
+        String url = OkapiUrlBuilder.getLogDeletionUrl(logUuid);
 
         final String responseBody = authedHttpGet(tokenProvider, url);